requirements.txt
psycopg2
Django
djangorestframework
django-rest-auth
djangorestframework-jwt
